+++ src/parse_cfg_file/laser_scan_cfg_parser.h
@@ -22,34 +22,28 @@
 // SOFTWARE.
 //
 
-// Denoting headers specifically used for building ROS1 Driver.
+#ifndef LIVOX_ROS_DRIVER_LASER_SCAN_CFG_PARSER_H_
+#define LIVOX_ROS_DRIVER_LASER_SCAN_CFG_PARSER_H_
 
-#ifndef ROS1_HEADERS_H_
-#define ROS1_HEADERS_H_
+#include "pointcloud_to_laserscan_converter.h"
+#include "rapidjson/document.h"
 
-#include <thread>
-#include <future>
+#include <string>
 
-#include <ros/ros.h>
-#include <rosbag/bag.h>
-#include <pcl_ros/point_cloud.h>
-#include <sensor_msgs/Imu.h>
-#include <sensor_msgs/LaserScan.h>
-#include <sensor_msgs/PointCloud2.h>
-#include "livox_ros_driver2/CustomMsg.h"
-#include "livox_ros_driver2/CustomPoint.h"
+namespace livox_ros {
 
+class LaserScanConfigParser {
+public:
+  static bool ParseLaserScanConfig(const rapidjson::Document &doc,
+                                   LaserScanConfig &scan_config);
+  static bool ParseLaserScanConfig(const std::string &config_file_path,
+                                   LaserScanConfig &scan_config);
 
-#define DRIVER_DEBUG(node, ...) ROS_DEBUG(__VA_ARGS__)
-#define DRIVER_INFO(node, ...) ROS_INFO(__VA_ARGS__)
-#define DRIVER_WARN(node, ...) ROS_WARN(__VA_ARGS__)
-#define DRIVER_ERROR(node, ...) ROS_ERROR(__VA_ARGS__)
-#define DRIVER_FATAL(node, ...) ROS_FATAL(__VA_ARGS__)
+private:
+  static bool ParseLaserScanFromValue(const rapidjson::Value &value,
+                                      LaserScanConfig &scan_config);
+};
 
-#define DRIVER_DEBUG_EXTRA(node, EXTRA, ...) ROS_DEBUG_##EXTRA(__VA_ARGS__)
-#define DRIVER_INFO_EXTRA(node, EXTRA, ...) ROS_INFO_##EXTRA(__VA_ARGS__)
-#define DRIVER_WARN_EXTRA(node, EXTRA, ...) ROS_WARN_##EXTRA(__VA_ARGS__)
-#define DRIVER_ERROR_EXTRA(node, EXTRA, ...) ROS_ERROR_##EXTRA(__VA_ARGS__)
-#define DRIVER_FATAL_EXTRA(node, EXTRA, ...) ROS_FATAL_##EXTRA(__VA_ARGS__)
+} // namespace livox_ros
 
-#endif // ROS1_HEADERS_H_
+#endif // LIVOX_ROS_DRIVER_LASER_SCAN_CFG_PARSER_H_